backend/test_docs/validation/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
typing-extensions>=4.0.0
google-search-results>=2.4.2
